+++ attributes/default.rb
@@ -190,11 +190,11 @@ default['is_apaas_openshift_cookbook']['openshift_master_console_url'] = "https:
 default['is_apaas_openshift_cookbook']['openshift_master_policy'] = "#{node['is_apaas_openshift_cookbook']['openshift_master_config_dir']}/policy.json"
 default['is_apaas_openshift_cookbook']['openshift_master_config_file'] = "#{node['is_apaas_openshift_cookbook']['openshift_master_config_dir']}/master-config.yaml"
 default['is_apaas_openshift_cookbook']['openshift_master_api_sysconfig'] = '/etc/sysconfig/atomic-openshift-master-api'
-default['is_apaas_openshift_cookbook']['openshift_master_api_systemd'] = '/usr/lib/systemd/system/atomic-openshift-master-api.service'
+default['is_apaas_openshift_cookbook']['openshift_master_api_systemd'] = '/etc/systemd/system/atomic-openshift-master-api.service'
 default['is_apaas_openshift_cookbook']['openshift_master_controllers_sysconfig'] = '/etc/sysconfig/atomic-openshift-master-controllers'
-default['is_apaas_openshift_cookbook']['openshift_master_controllers_systemd'] = '/usr/lib/systemd/system/atomic-openshift-master-controllers.service'
+default['is_apaas_openshift_cookbook']['openshift_master_controllers_systemd'] = '/etc/systemd/system/atomic-openshift-master-controllers.service'
 default['is_apaas_openshift_cookbook']['openshift_master_ca_certificate'] = { 'data_bag_name' => nil, 'data_bag_item_name' => nil, 'secret_file' => nil }
-default['is_apaas_openshift_cookbook']['openshift_master_named_certificates'] = %w[]
+default['is_apaas_openshift_cookbook']['openshift_master_named_certificates'] = node['is_apaas_openshift_cookbook']['openshift_node_user_data'] && JSON.parse(node['ec2']['userdata']).key?('openshift_master_named_certificates') ? JSON.parse(node['ec2']['userdata'])['openshift_master_named_certificates'] : %w[]
 default['is_apaas_openshift_cookbook']['openshift_master_scheduler_conf'] = "#{node['is_apaas_openshift_cookbook']['openshift_master_config_dir']}/scheduler.json"
 default['is_apaas_openshift_cookbook']['openshift_master_managed_names_additional'] = %w[]
 default['is_apaas_openshift_cookbook']['openshift_master_retain_events'] = nil
@@ -245,8 +245,8 @@ default['is_apaas_openshift_cookbook']['openshift_hosted_router_deploy_shards'] 
 default['is_apaas_openshift_cookbook']['openshift_hosted_router_shard'] = []
 default['is_apaas_openshift_cookbook']['openshift_hosted_router_namespace'] = 'default'
 default['is_apaas_openshift_cookbook']['openshift_hosted_router_options'] = []
-default['is_apaas_openshift_cookbook']['openshift_hosted_router_certfile'] = "#{node['is_apaas_openshift_cookbook']['openshift_master_config_dir']}/openshift-router.crt"
-default['is_apaas_openshift_cookbook']['openshift_hosted_router_keyfile'] = "#{node['is_apaas_openshift_cookbook']['openshift_master_config_dir']}/openshift-router.key"
+default['is_apaas_openshift_cookbook']['openshift_hosted_router_certfile'] = node['is_apaas_openshift_cookbook']['openshift_node_user_data'] && JSON.parse(node['ec2']['userdata']).key?('openshift_hosted_router_certfile') ? JSON.parse(node['ec2']['userdata'])['openshift_hosted_router_certfile'] : "#{node['is_apaas_openshift_cookbook']['openshift_master_config_dir']}/openshift-router.crt"
+default['is_apaas_openshift_cookbook']['openshift_hosted_router_keyfile'] = node['is_apaas_openshift_cookbook']['openshift_node_user_data'] && JSON.parse(node['ec2']['userdata']).key?('openshift_hosted_router_keyfile') ? JSON.parse(node['ec2']['userdata'])['openshift_hosted_router_keyfile'] : "#{node['is_apaas_openshift_cookbook']['openshift_master_config_dir']}/openshift-router.key"
 
 default['is_apaas_openshift_cookbook']['openshift_hosted_manage_registry'] = true
 default['is_apaas_openshift_cookbook']['openshift_hosted_registry_namespace'] = 'default'