+++ recipes/node.rb
@@ -110,7 +110,19 @@ if is_node_server
     version Array.new(pkg_node_array.size, node['is_apaas_openshift_cookbook']['ose_version']) unless node['is_apaas_openshift_cookbook']['ose_version'].nil?
     options node['is_apaas_openshift_cookbook']['openshift_yum_options'] unless node['is_apaas_openshift_cookbook']['openshift_yum_options'].nil?
     not_if { node['is_apaas_openshift_cookbook']['deploy_containerized'] }
+    flush_cache [:before]
     retries 3
+  end
+
+  if node['is_apaas_openshift_cookbook']['asynchronous_upgrade']
+    log 'Node services (Asynchronous Upgrade)' do
+      level :info
+      notifies :restart, 'service[Restart Node]', :immediately
+      not_if { ::File.file?("/usr/local/share/info/.upgrade-#{node['is_apaas_openshift_cookbook']['ose_version']}") }
+    end
+    file "/usr/local/share/info/.upgrade-#{node['is_apaas_openshift_cookbook']['ose_version']}" do
+      action :create_if_missing
+    end
   end
 
   yum_package 'conntrack-tools' do